requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv